+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Changelog ##
 #### dev ####
 * Update how workshop info is saved and give mods access to it
+* Add flag for selecting specific mods
+  * Implies --skip-launcher
 
 #### v3.24.0 ####
 * Change how out-jar backend works